# backend/examples/webflux-liquibase/app/src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/
    description: Generated server url
tags:
  - name: car
    description: Manage Cars
paths:
  /api/cars/{id}:
    get:
      tags:
        - car
      summary: car controller
      description: add cars
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CarDto"
  /api/cars/byName:
    get:
      tags:
        - car
      operationId: findByName
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            "$ref": "#/components/schemas/Pageable"
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/PageCarEntity"
components:
  schemas:
    CarDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        carName:
          type: string
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    CarEntity:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
    PageCarEntity:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        content:
          type: array
          items:
            "$ref": "#/components/schemas/CarEntity"
        sort:
          "$ref": "#/components/schemas/Sort"
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          "$ref": "#/components/schemas/Pageable"
        empty:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean